+++ app/src/main/java/opensource/cached_dupe_scanner/MainActivity.kt
@@ -259,6 +259,7 @@ class MainActivity : ComponentActivity() {
                                 keepLoadedThumbnailsInMemory = settingsSnapshot.keepLoadedThumbnailsInMemory,
                                 keepLoadedVideoPreviewsInMemory = settingsSnapshot.keepLoadedVideoPreviewsInMemory,
                                 snapVideoPreviewFramesToWidth = settingsSnapshot.snapVideoPreviewFramesToWidth,
+                                videoPreviewLineCount = settingsSnapshot.videoPreviewLineCount,
                                 thumbnailSizeScale = settingsSnapshot.thumbnailSizePercent / 100f,
                                 videoPreviewSizeScale = settingsSnapshot.videoPreviewSizePercent / 100f,
                                 rememberedThumbnailCache = rememberedThumbnailCache,